Use int for port
git-svn-id: svn+ssh://subversion.munts.net/src/trunk/libsimpleio@13063 06ee2f96-5407-11de-8bab-f3d7589a4122

+++ java/libtcp4.java
@@ -38,13 +38,13 @@ public class libtcp4
   public static native void TCP4_ntoa(int addr, byte[] dst, int dstsize,
     IntByReference error);
 
-  public static native void TCP4_connect(int addr, short port,
+  public static native void TCP4_connect(int addr, int port,
     IntByReference fd, IntByReference error);
 
-  public static native void TCP4_accept(int addr, short port,
+  public static native void TCP4_accept(int addr, int port,
     IntByReference fd, IntByReference error);
 
-  public static native void TCP4_server(int addr, short port,
+  public static native void TCP4_server(int addr, int port,
     IntByReference fd, IntByReference error);
 
   public static native void TCP4_send(int fd, byte[] buf, int bufsize,
